[test] optional RED_TEST_DELEGATE_PRINT_NS for global enum. add RED_TEST_DELEGATE_PRINT_ENUM_NS(ns, type)

+++ tests/core/RDP/capabilities/test_surfacecommands.cpp
@@ -32,7 +32,7 @@ RED_AUTO_TEST_CASE(TestCapabilitySurfaceCommandsEmit)
     surfacecommands_caps.cmdFlags = 65536;
     surfacecommands_caps.reserved = 65536;
 
-    RED_CHECK_EQUAL(surfacecommands_caps.capabilityType, static_cast<uint16_t>(CAPSETTYPE_SURFACE_COMMANDS));
+    RED_CHECK_EQUAL(surfacecommands_caps.capabilityType, CAPSETTYPE_SURFACE_COMMANDS);
     RED_CHECK_EQUAL(surfacecommands_caps.len, static_cast<uint16_t>(CAPLEN_SURFACE_COMMANDS));
     RED_CHECK_EQUAL(surfacecommands_caps.cmdFlags, static_cast<uint32_t>(65536));
     RED_CHECK_EQUAL(surfacecommands_caps.reserved, static_cast<uint32_t>(65536));
@@ -44,7 +44,7 @@ RED_AUTO_TEST_CASE(TestCapabilitySurfaceCommandsEmit)
 
     SurfaceCommandsCaps surfacecommands_caps2;
 
-    RED_CHECK_EQUAL(surfacecommands_caps2.capabilityType, static_cast<uint16_t>(CAPSETTYPE_SURFACE_COMMANDS));
+    RED_CHECK_EQUAL(surfacecommands_caps2.capabilityType, CAPSETTYPE_SURFACE_COMMANDS);
     RED_CHECK_EQUAL(surfacecommands_caps2.len, static_cast<uint16_t>(CAPLEN_SURFACE_COMMANDS));
 
     RED_CHECK_EQUAL(static_cast<uint16_t>(CAPSETTYPE_SURFACE_COMMANDS), stream.in_uint16_le());
